more stuff

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 //@desc     Get goals
 //@route    GET /api/goals
 //@access   Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({user: req.user.id});
 
     res.status(200).json(goals);
 });
@@ -22,6 +23,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     })
 
     res.status(201).json(goal);
@@ -36,6 +38,11 @@ const updateGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(404);
         throw new Error('Goal not found');
+    }
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true});
@@ -53,7 +60,12 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error('Goal not found');
     }
-    
+
+    if(goal.user.toString() !== req.user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
+    }
+
     await Goal.findByIdAndDelete(req.params.id);
 
     res.status(200).json({id: req.params.id});
